Remove data generator duplicate.
