+++ src/benchmarks/micro/corefx/System.Text.Json/Serializer/Json_FromStream.cs
@@ -8,6 +8,7 @@ using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
 using MicroBenchmarks;
+using MicroBenchmarks.Serializers;
 
 namespace System.Text.Json.Tests
 {
